[DLCov] Annotate intentionally dropped/empty DebugLocs in existing code
Following the work in PR #107279, this patch applies the annotative
DebugLocs, which indicate that a particular instruction is intentionally
missing a location for a given reason, to existing sites in the compiler
where their conditions apply. This is a no-op in ordinary LLVM builds (each
function `get<Type>` simply expands to `DebugLoc()`), but marks the
instruction in coverage-tracking builds so that it will be ignored by
Debugify, allowing only real errors to be reported. As a side-effect of
this, it also clarifies the intention behind the missing DebugLoc to a
reader.